// dto/stationEdit.dto.ts
export interface StationEditDto {
  nome: string
  descricao: string
  horarioFuncionamento: string
  tipoTomada: string
  comodidade: string
  statusFuncionamento: string
  precoKwh: number
  cabo: number
  potencia: number
  latitude: number
  longitude: number
  endereco: string
  estado: string
  cep: string
  cidade: string
  numero: number
  complemento: string
}